+++ mlir-sycl/lib/Dialect/SYCL/Analysis/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_mlir_dialect_library(MLIRSYCLAnalysis
   AliasAnalysis.cpp
   MemoryAccessAnalysis.cpp  
-  ReachingDefinitionAnalysis.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_SYCL_INCLUDE_DIR}/mlir/Analysis
