+++ mlir-sycl/test/CMakeLists.txt
@@ -1,5 +1,3 @@
-add_subdirectory(lib)
-
 configure_lit_site_cfg(
   ${CMAKE_CURRENT_SOURCE_DIR}/lit.site.cfg.py.in
   ${CMAKE_CURRENT_BINARY_DIR}/lit.site.cfg.py
