+++ mlir-sycl/tools/sycl-mlir-opt/CMakeLists.txt
@@ -1,7 +1,3 @@
-set(test_libs
-  SYCLMLIRTestAnalysis
-)
-
 set(LIBS
   ${test_libs}
   MLIRMlirOptMain
