+++ polygeist/include/mlir/Dialect/Polygeist/Analysis/ReachingDefinitionAnalysis.h
@@ -10,50 +10,14 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef MLIR_DIALECT_SYCL_ANALYSIS_REACHINGDEFINITIONANALYSIS_H
-#define MLIR_DIALECT_SYCL_ANALYSIS_REACHINGDEFINITIONANALYSIS_H
+#ifndef MLIR_DIALECT_POLYGEIST_ANALYSIS_REACHINGDEFINITIONANALYSIS_H
+#define MLIR_DIALECT_POLYGEIST_ANALYSIS_REACHINGDEFINITIONANALYSIS_H
 
 #include "mlir/Analysis/DataFlow/DenseAnalysis.h"
-#include "mlir/IR/FunctionInterfaces.h"
+#include "mlir/Dialect/Polygeist/Utils/AliasUtils.h"
 
 namespace mlir {
-class AliasAnalysis;
-namespace sycl {
-
-class AliasUtilities {
-  friend raw_ostream &operator<<(raw_ostream &, const AliasUtilities &);
-
-public:
-  // Construct alias utilities for function \p funcOp.
-  AliasUtilities(FunctionOpInterface &funcOp,
-                 mlir::AliasAnalysis &aliasAnalysis);
-
-  /// Return the set of values that are definitely aliased to \p val.
-  SetVector<Value> getMustAlias(Value val) const {
-    return valueToMustAliasValues.lookup(val);
-  }
-
-  /// Return the set of values that are possibly aliased to \p val.
-  SetVector<Value> getMayAlias(Value val) const {
-    return valueToMayAliasValues.lookup(val);
-  }
-
-  mlir::AliasAnalysis &getAliasAnalysis() const { return aliasAnalysis; }
-
-private:
-  /// Collect values that reference a memory resource within function \p funcOp.
-  SetVector<Value> collectMemoryResourcesIn(FunctionOpInterface funcOp);
-
-private:
-  // val -> values that are definitely aliased.
-  DenseMap<Value, SetVector<Value>> valueToMustAliasValues;
-
-  // val -> values that might be aliased.
-  DenseMap<Value, SetVector<Value>> valueToMayAliasValues;
-
-  FunctionOpInterface &funcOp;
-  mlir::AliasAnalysis &aliasAnalysis;
-};
+namespace polygeist {
 
 /// This lattice represents the set of operations that might have modified a
 /// memory resource last.
@@ -139,11 +103,11 @@ public:
   void setToEntryState(ReachingDefinition *lattice) override;
 
 private:
-  DenseMap<FunctionOpInterface, std::unique_ptr<AliasUtilities>> aliasUtilities;
+  DenseMap<FunctionOpInterface, std::unique_ptr<AliasQueries>> aliasQueriesMap;
   mlir::AliasAnalysis &aliasAnalysis;
 };
 
-} // namespace sycl
+} // namespace polygeist
 } // namespace mlir
 
-#endif // MLIR_DIALECT_SYCL_ANALYSIS_REACHINGDEFINITIONANALYSIS_H
+#endif // MLIR_DIALECT_POLYGEIST_ANALYSIS_REACHINGDEFINITIONANALYSIS_H
