+++ polygeist/lib/Dialect/Polygeist/Utils/AliasUtils.cpp
@@ -1,0 +1,91 @@
+//===- AliasUtils.cpp - Alias Queries Utilities  --------------------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "mlir/Dialect/Polygeist/Utils/AliasUtils.h"
+#include "mlir/Analysis/AliasAnalysis.h"
+#include "mlir/Interfaces/SideEffectInterfaces.h"
+#include "llvm/Support/raw_ostream.h"
+
+namespace mlir {
+namespace polygeist {
+
+//===----------------------------------------------------------------------===//
+// AliasQueries
+//===----------------------------------------------------------------------===//
+
+raw_ostream &operator<<(raw_ostream &os, const AliasQueries &aliasQueries) {
+  auto printMap = [&os](const DenseMap<Value, SetVector<Value>> &map,
+                        StringRef title) {
+    for (const auto &entry : map) {
+      Value val = entry.first;
+      const SetVector<Value> &mustAlias = entry.second;
+      os.indent(4) << val << "\n";
+      os.indent(4) << title << ":\n";
+      if (mustAlias.empty())
+        os.indent(6) << "<none>\n";
+      else {
+        for (Value aliasedVal : mustAlias)
+          os.indent(6) << aliasedVal << "\n";
+      }
+      os << "\n";
+    }
+  };
+
+  os.indent(4) << aliasQueries.funcOp.getName() << ":\n";
+  printMap(aliasQueries.valueToMustAliasValues, "mustAliases");
+  printMap(aliasQueries.valueToMayAliasValues, "mayAliases");
+  return os;
+}
+
+AliasQueries::AliasQueries(FunctionOpInterface &funcOp,
+                           mlir::AliasAnalysis &aliasAnalysis)
+    : funcOp(funcOp), aliasAnalysis(aliasAnalysis) {
+  /// Collect all operations that reference a memory resource in the given
+  /// function and initialize the maps.
+  SetVector<Value> memoryResources = collectMemoryResourcesIn(funcOp);
+  for (Value val1 : memoryResources) {
+    for (Value val2 : memoryResources) {
+      if (val1 == val2)
+        continue;
+
+      AliasResult aliasResult = aliasAnalysis.alias(val1, val2);
+      if (aliasResult.isMust())
+        valueToMustAliasValues[val1].insert(val2);
+      else if (aliasResult.isMay())
+        valueToMayAliasValues[val1].insert(val2);
+    }
+  }
+}
+
+SetVector<Value>
+AliasQueries::collectMemoryResourcesIn(FunctionOpInterface funcOp) {
+  SetVector<Value> memoryResources;
+  for (BlockArgument arg : funcOp.getArguments()) {
+    if (isa<MemRefType>(arg.getType()))
+      memoryResources.insert(arg);
+  }
+
+  for (Block &block : funcOp) {
+    for (Operation &op : block.without_terminator()) {
+      auto memoryEffectOp = dyn_cast<MemoryEffectOpInterface>(op);
+      if (!memoryEffectOp)
+        continue;
+
+      SmallVector<MemoryEffects::EffectInstance> effects;
+      memoryEffectOp.getEffects(effects);
+      for (const auto &effect : effects) {
+        if (Value val = effect.getValue())
+          memoryResources.insert(val);
+      }
+    }
+  }
+  return memoryResources;
+}
+
+} // namespace polygeist
+} // namespace mlir
