+++ polygeist/lib/Dialect/Polygeist/Utils/CMakeLists.txt
@@ -1,13 +1,15 @@
 add_mlir_dialect_library(MLIRPolygeistUtils
+  AliasUtils.cpp
   TransformUtils.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_MAIN_INCLUDE_DIR}/mlir/Dialect/Polygeist/Utils
 
   LINK_LIBS PUBLIC
+  MLIRAnalysis
+  MLIRDialect  
   MLIRGPUOps
+  MLIRIR  
   MLIRPolygeistDialect
   MLIRSYCLDialect
-  MLIRDialect
-  MLIRIR
   )
